+++ benchmark/poly1305.dart
@@ -1,0 +1,37 @@
+// Copyright (c) 2021, Sudipto Chandra
+// All rights reserved. Check LICENSE file for details.
+
+import 'dart:math';
+
+import 'package:hashlib/hashlib.dart' as hashlib;
+
+import 'base.dart';
+
+Random random = Random();
+
+final key = List.generate(16, (i) => random.nextInt(256));
+final secret = List.generate(16, (i) => random.nextInt(256));
+
+class HashlibBenchmark extends Benchmark {
+  HashlibBenchmark(int size, int iter) : super('hashlib', size, iter);
+
+  @override
+  void run() {
+    hashlib.poly1305(input, key, secret);
+  }
+}
+
+void main() {
+  print('------- Poly1305 --------');
+  final conditions = [
+    [10, 100000],
+    [1000, 5000],
+    [500000, 10],
+  ];
+  for (var condition in conditions) {
+    int size = condition[0];
+    int iter = condition[1];
+    HashlibBenchmark(size, iter).showDiff();
+    print('');
+  }
+}