+++ lib/src/hashlib_base.dart
@@ -15,6 +15,7 @@ export 'keccak384.dart';
 export 'keccak512.dart';
 export 'md5.dart';
 export 'pbkdf2.dart';
+export 'poly1305.dart';
 export 'scrypt.dart';
 export 'sha1.dart';
 export 'sha224.dart';